stepik_brute-forcer/requirements.txt
# Основные зависимости
playwright>=1.40.0
cryptography>=41.0.0
schedule>=1.2.0
websocket-client>=1.6.0
numpy>=1.24.0
pandas>=2.0.0
python-dateutil>=2.8.0

# Для GUI (Вариант 2 - ttkthemes + Pillow)
ttkthemes>=3.2.2
pillow>=10.0.0

# Для разработки и тестирования
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-playwright>=0.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0